Add BSD+Patent license in ReST format
The Open Source Initiative (OSI) approved BSD+Patent license is
essentially the BSD 2-clause license with language that grants
users the additional right to keep using the software even if
some or all of it is patented by contributors.  The list of
conditions in the license does not require that the patent grant
be included in the PlasmaSimulator license when it is
redistributed.